Reunited ED_KANE_MELE_AFMz
After some testing on full-diagrams has been observed that:

- The nonsu2-mode/replica-bath implementation of the AFMz calculation is indeed far easier to converge, at a fixed DMFT_ERROR (i.e. does not require small mixing parameters nor U-steps, even at the transition...which seems too much, to be fair)

- The resulting converged data is nevertheless far dirtier wrt the normal-mode/normal-bath calculation, making all the advantage vanish. In particular: the Uc-line resulted to be very bad (even loosing its monotonicity!); Initial AFMkick appears to be insufficient to determine the sign of MagZ (!).

All of this may hint to some problem on one or both the implementations. Since the AMFz_normal case has been succesfully tested (in the zero SO coupling case) against PRB 101 125103 (2020), I'd say the prior is that the AFMz_replica implementation is either wrong or inappropriated wrt the problem, for some reason. More testing may shade light on the issue.

SO, FOR THE TIME BEING: the AFMz_replica implementation will be merged with the AFMz_normal one, in a unique AFMz driver. Since the two cases share large part of the input state and actual implementation at this point I deem pleonastic to keep two separate drivers. Of course if the _replica case was to be proved unuseful the driver will be reverted to the original normal-mode/normal-bath state.