Update RunningDMFT_refresh.m
Added control variable for max number of dmft loops in the refresh. THIS IS PARAMOUNT!

>> Some context and motivation: lenghty style 'cause I need to reflect quite a bit on this yet.

When you resume a dmft calculation from the .restart files, you are actually resuming from an instantaneous state, loosing track of the *previous loops history* (which has a nonzero correlation time due to self-mixing of the solution!). Think about the whole dmft procedure as a minimization process in some parameter space (like every self-consistent approach...). If the minimum is "easy-to-reach" you'll get _arbitrary close_ to it in a few steps and the dmft_error will go down very fast. Whatever definition of this error will get you again very fast to very small values, even if you restart loosing your "path-history", since you will be for sure already considerably close to the true minimum, in parameter space.
If instead you have a very flat landscape and a hard search for the minimum you will have that any reasonable definition of convergence will lead to a relatively large neighbourhood of the true minimum, in parameter space. So that an "arbitrary small" distance from the exact solution, in the "self-consistency-stability-space" would correspond to a comparatively not-so-small distance in parameter space: you are physically close to the solution, but you are not so close to the exact parameters identifying it. Hence *if and only if* you have track of your iterative path you can determine that you are close to self-consistency, cause you are getting arbitrary near your "candidate minimum" at each new loop. But if you restart resetting any past information, you will very possibily make a huge leap inside this "big neighbourhood" and then spend quite a lot of loops before again seemingly converge, i.e. having small steps in parameter space, due to a new "self-mixing-aided" stabilization of the path. But if you were from the (re)start already inside the "converged neighbourhood" it has no sense to loop a lot again, since the old _converged_ candidate and the new _re-converged_ one will be on average equally distant from the true minimum of the problem.  Ideally if you just refresh with just one loop, and you restart from hamiltonian.used instead of hamiltonian.restart, you have to get the very same observables and GFs, at machine precision, then the old calculation. But you will virtually get a 1e00 error for you have no reference to which compute a distance! Nevertheless the quality of the solution obviously remains the same (e.g. the old calculation had probably ended with a very good <e-05 dmft_error). 

So definitely, we won't tipically want to loop again until __formal__ convergence, but just refresh with a quite small number of new dmft loops (ideally 1 is already ok, for refreshing a converged solution). Nevertheless I leave the option to increase Nnew a bit (let's say up to 10?) if the refresh is motivated by bad convergence of the old calculation and you really want to improve it a bit (e.g. noisy observables).

>> To Do

It might be quite useful to move all the hamiltonian.used into the correspective hamiltonian.restart files, within the script. So that, e.g. if Nnew=0, we could actually force a machine-precision-exact reproduction of the already stored data: quite desirable for reproducibility (e.g. we want to refresh the calculation to correct a bug affecting some observable, but want to mantain the very same data for the unaffected remaining ones).

But I'll need to think a bit about it, since I don't want to mess with the inequivalent lattices suffixes and screw up the restart condition.