Update KMH-DMFT drivers
Mirror of commit < https://github.com/QcmPlab/Driver-Database/commit/e3681bdf4cc790e34588c6a663bae7e51715ee29 >

Note that in the meanwhile a most relevant bug has been fixed in LIB_DMFT_ED: < https://github.com/QcmPlab/LIB_DMFT_ED/commit/0e5c272b45eda6b7ff652e2473b9ecda09e5ba8b >

So currently we have:

=== SOI=0, U=2t -> unbroken phase ====================
BATHx and BATHz give very similar observables (but not equal to machine precision), convergence is far slower for BATHz, so that's I believe why we get such a numerical difference in computed quantities: loop by loop the AIM parameters evolve
arm in arm, to real machine precision and so do the observables... but the BATHz solution ends up "very far" for it perform a hundred more loops!
================================================

=== SOI=0, U=10t -> far in the AFM phase ================
BATHx and BATHz converge both very fast (about 5 loops); AIM, order parameter, observables and energies match to machine precision, as it should -> no SOI, no frustration -> full SU(2) sym
================================================

=== SOI=0.1t, U=6t&10t -> AFM phase ===================
BATHx and BATHz give order parameter, observables and energies equal up to a 1% error. In both cases Egs(AFMx) < Egs(AFMz), but the energy difference, if normalized to the relevant scale (U) is very very tight with respect to the relative difference of the order parameters... so is it even there, or is it just a numerical artefact?
At least we have Mz > Mx, so it's not just that we get more energy from just having "more order parameter". Still I doubt we can really claim a proof of favoured in-plane magnetism with these kind of results...
================================================

Finally, this recalls an ancient problem with the MF energies (cfr. commit b0017ab95c54c3bb9919743fe122c2fcdc795fe1), namely that we have not the same normalization **convention** on the two setups: DMFT has a basically constant energy profile in the unbroken phase whereas MF has a term that is **linear in U**, so that there is NO WAY to enforce the same zero-energy at each U-point. We NEED to solve this issue if we want to really compare the MF and DMFT answers, and we definitely do, for the MF has a way stronger preference for the AFMx solution... so the key to solve the whole thing should be hidden here.