Fix MF | Part II
Looking at the MF bands... noninteracting limit appears to be correct, if
compared to Eur. Phys. J. B 68, 537–542 (Jafari, 2009). [As a side note,
this is in contrast to what GFtools reports for the honeycomb lattice, we
might want to consider this in MOTTlab... is it D=3/2t or D=3t? Damn...]

- We change the default T1 in the F90 source, to avoid any future problem.

- We also change the number of k-points per direction, so to be attuned
  with the DMFT defaults. Maybe this alone could make for the discrepancy.
  > I find it very doubtful, but who knows.