Fix MF | Part IV
I'm so, so dumb. I was summing on the BZ-PATH to get the mean-field band
contribution to the ground-state. This is really fucking dumb.

So now we sum over THE WHOLE BZ within the FORTRAN driver, and it works!
> the noninteracting limit of Emf now matches exactly with the DMFT code

BUT

The strong coupling limit now is broken: MF energy goes under DMFT result!
> The MB term normalization is actually off, I'd guess commit 3a54794f564d
  represents the perfect case of confirmation bias.

* S A D N E S S *