Add a SLURM-array aware script for DMFT 2D phase diagrams
It reads the array task id via getenv(), and converts the result to double.
If the resulting number is not a NaN (i.e. if the env variable is defined),
it avoid looping over all spin-orbit coupling values and just computes the
line selected by slurm.

The called fixed-SO lines are of 'dry' type, in the runDMFT sense.