FIX: Hartree-Fock self-energy [Issue #3]
The self-energy evaluation introduced within commit cc0f13b has been found
to be severely flawed, despite leading to a reasonable kinetic energy. In
both the real and imaginary axes the resulting ∑(z) is /not/ causal, nor
in any sense constant.

After some rounds of debugging (here we add many temporary printings...)
I moved to MATLAB to see if a corresponding implementation of the Dyson
would lead to the same flawed spectra, or what. What happened instead is
that MATLAB readily complaints about bad-conditioned matrix-inversions,
which after all I could totally foresee, given that G(ω) is disseminated
with zeros and G(iω) has exponentially vanishing tails.

More specifically we get, for almost all matsubara frequencies and quite
a lot real ones:

>> Warning: Matrix is close to singular or badly scaled.
   Results may be inaccurate. RCOND = xxxxxxxx.

where typical values of RCOND fall in the (E-20,E-16) range.
> https://it.mathworks.com/help/matlab/ref/rcond.html

SciFortran does not complain in any way, but I see no reason for the two
implementations to behave differently regarding this kind of issues.

————————————————————————————————————————————————————————————————————————

Hence here we remove the Dyson's equation and build the self-energy in a
direct way:

\Sigma_HF = Hmf_correction

from the following (admittedly convoluted) reasoning:

- HF model is both interacting and single particle, hence we have a
  trivial constant self-energy \Sigma(∞) = \Sigma(0) = \Sigma(ω) ∀ω

- Its topological renormalization is defined in terms of \Sigma(0) as
  Htop = H0 + \Sigma(0) = H0 + \Sigma(∞) = H0 + \Sigma(ω) ∀ω

- Being already a single-particle model, its topological hamiltonian
  should match the model itself, thus Htop = Hmf = H0 + Hmf_correction

==> All these imply that ∀ω \Sigma(ω) = Hmf_correction.

Probably there is a better way to prove this, for the generic complex
frequency z, by explicitly deriving the Hartree-Fock renormalization of
the single particle levels, but at the moment I do not have it.

————————————————————————————————————————————————————————————————————————

Trusting the build on both axes we get:

- correct constant self-energies, in both the AFMxy and AFMz decouplings

- correctly vanishing spin-diagonal self-energy terms in the AFMxy case

- correctly vanishing spin-off-diagonal terms in the AFMz solution

- a kinetic energy with the same qualitative behavior, but with a better
  match at both weak and strong coupling with the corresponding DMFT.

How could it be that the kinetic energy was "not so much wrong"?

> inspecting the old AFMz self-energies, on the matsubara axis, I find
  that the imaginary part was totally off (not even causal), but the
  real part was actually slow varying (at low frequency), with a infty
  value very near to the new correct prediction. So in some cursed sense
  we were almost there... oh my numerics! 🤣