Update `ed_km_flakes`
There was a serious problem in the hamiltonian constructions: in the km
model the next-nearest-neighbors hopping need not only the information
about the shell, but also the "direction" of the hopping, i.e. if we are
traveling the hexagon clock-wise or anti-clock-wise. This is impossible
to get from a type(xy_lattice) object for that no particular ordering is
guaranteed. We need instead to build again the type(hex) array and build
the SOC terms for each hexagon (where the ordering is clear and fixed..)
and then do a xy_lattice lookup to assign the terms to the correct (i,j)
element. This implies to import many many more modules from HoneyTools.

Preliminary testing is promising, but I need to crosscheck with reliable
results: for that I would prepare a "supercell" version of the driver to
be compared to the standard haldane-obc program given in the PythTB docs.
> http://www.physics.rutgers.edu/pythtb/examples/edge.py