Add `ed_km_supercell.f90` and `tb_qhi_supercell.py`
Unfortunately they cannot really be compared, since PythTB does build the
supercell with a slightly different convention than ours: they end up with
2 more atoms at the corners. Hence the comparison can be carried only on a
qualitative basis.

Mh=t1=1, t2=0, Bz=0
———————————————————
We have a perfect qualitative match on the single-particle groundstate,
but less on the "edge states": the lack of the single atom at the corner
makes for a great difference (in PythTB it takes almost all density, while
for us it is splitted in the _two_ corner atoms). The energies cannot be
compared one by one, but overall there is a _fair_ match, with greatest
difference at the Fermi level (reasonable...).

t1=1, Mh=0, t2=0.1, Bz=0
————————————————————————
Here the greatest problems, the edge states initially were totally wrong,
but apparently we write down the Hup and Hdw matrices with low precision,
from fortran, since the diagonalization of Hup within matlab and the same
eigenvectors as printed by Fortran differ //qualitatively//! Hence I just
use the fortran generated eigenvectors and at least obtain an edge-state
character, though not really correspondent to the PythTB plot: it breaks
mirror symmetry along a plane, for example.

We should find a way to test more precisely all of this.