This is not really fixed but hints at what is the problem
Apparently the k points spanning the surface are generated in reduced
coordinates. Still a lot of problems so there should be still something
missing on my side. I'm tired of all this so for now I'd rather stick to
their implementation for the Haldane spin hamiltonian, which we know to
work just fine. Hopefully we would not find trouble when mixing with our
Σ(0) from the DMFT (which defines the KM hamiltonian with our convention).