More or less working implementation of the KM model with `Z2Pack`
Still some doubt on the results around the topological transition, as in
Mh = 1.6t1, t2 = 0.55t1. We get a trivial insulator but it should still
be trivial (Mh does not exceeds 3*sqrt(3)*t2 yet). Careful inspection of
the hybrid Wannier center evolution shows that it is not truly a problem
of Z2Pack's algorithm to compute the Wilson loop but rather a issue with
computing the Wannier center evolution itself, as it differs from PythTB.
Visual inspection of PythTB's hybrid Wannier centers in fact determines
a correct Z2 invariant. So what's still missing in the model? Do we have
an error somewhere (but the haldane model is from their examples...) or
there is just an inherent imprecision in the kdotp approximation?

NB. The problem is not just the Z2 invariant for the Kane Mele model but
    it arises also in the spin-Chern numbers for the single Haldane's Hk.
    In fact the Z2 always matches (Cup-Cdw)/2. Hence it is not something
    we are missing in combining HkUP and HkDW, but it should hint to an
    issue in the Haldane model already.