Topological Hamiltonian in KM flakes (+ no more NEELSYM, please)
Since flakes are really messy and there's a lot of files, I deem much
better to build the topological Hamiltonian within the Fortran driver,
where Smats is available in the scope as the usual multi-dimensional
[Nlat,Nspin,Nspin,Norb,Norb,Lmats] array. Hence we would just have:

   Htop = kmHij + nnn2lso_reshape(Smats(:,:,:,:,:,1),Nlat,Nspin,Norb)
   call TB_write_Hloc(Htop,'topoHij.dat')

to obtain a nice TB_IO formatted matrix for the topological Hamiltonian.

NOTE: here I also delete all the misleading treatment of Neel symmetry.