Add `strawberrypy` scripts for computing the Z2 real-space marker
run/pre_strawberry prepares the files to feed the temporary version of
                   strawberrypy that Nicolas Baù kindly prepared for us.

run/run_strawberry runs the python code to compute the Z2 marker from
                   each topological hamiltonian from RDMFT and saving
                   appropriate txt and pdf files.

lib/strawberrypy_tmp stores the temporary version of strawberrypy we are
                     using here.

Briefly tested on brewed macOS on three selected points. Let's see if we
can run this on ulysses to process the actual RDMFT data.