Revert "Let's try deactivating the macroscopic average"
This reverts commit 7b5aed7dbe6709fc982d8ef1c5dd78a47711c1d8.

If we pass "False" to macroscopic_average keyarg we fallback to a unit
cell average, computed with an implicit assumption on the indexing of
the sites (the indexing of pythtb). The correct way to look at the raw
real-space operator (no average whatsoever) is to set the cutoff to 0.

We also change (temporarily) the colormap, so to inspect with more care
the convergence of the bulk value.